requirements.txt
aiogram
yt-dlp
shazamio
pandas
openpyxl
python-dotenv
aiofiles